+++ csum-file.h
@@ -9,7 +9,7 @@ struct progress;
 struct hashfile {
 	int fd;
 	int check_fd;
-	unsigned int offset;
+	size_t offset;
 	git_hash_ctx ctx;
 	off_t total;
 	struct progress *tp;
@@ -37,7 +37,7 @@ struct hashfile *hashfd(int fd, const char *name);
 struct hashfile *hashfd_check(const char *name);
 struct hashfile *hashfd_throughput(int fd, const char *name, struct progress *tp);
 int finalize_hashfile(struct hashfile *, unsigned char *, unsigned int);
-void hashwrite(struct hashfile *, const void *, unsigned int);
+void hashwrite(struct hashfile *, const void *, size_t);
 void hashflush(struct hashfile *f);
 void crc32_begin(struct hashfile *);
 uint32_t crc32_end(struct hashfile *);
